Hide quote of original message
When answering a carteggio message, normal email clients quote the
original text. This commit hides the quote in the conversation view and
instead shows three small dots at the end of the message. If the user
wants to see the full message with the quote he must tap on the message.